Add test for the current behavior of the parser
The test cases are mostly based on the following resources:

https://nixos.org/guides/nix-pills/basics-of-language.html
https://nixos.org/manual/nix/stable/#ch-expression-language

Note that the outcome expected by the tests is not based on the behavior
described in the mentioned resources but on the current behavior of the
parser. The expected outcome will be corrected in upcoming commits when
the related bugs of the parser are fixed. This approach shall make
bugfixes on the parser easier to review by showing the old and new
behavior in the diff.