Disable download of IntelliJ sources on build server
The sources of IntelliJ can be usedful during development in order to
understand the API or search for examples of similar use cases. However,
it does not have any value on the build server but increases the build
time and cache size. By using the default configuration of the
gradle-intellij-plugin, sources are only downloaded if the CI
environment variable is *not* set. GitHub Actions already sets the CI
environment variable by default.