Fix GitHub Actions running out of disk space
The build in GitHub Actions was running out of disk while executing the
post/cleanup-actions. This sometimes caused an exception inside the code
of the runner:

  System.IO.IOException: No space left on device

Most disk space was occupied by different versions of IDEA. Most of them
were downloaded by :runPluginVerifier. By running the task for each
version of IDEA in a separate job, we should be able to avoid running
out of disk space.