Consolidate token set inside NixTokenSets
Note that before this commit, there was a warning from IntelliJ that I
should not access NixTypes within the static initializer (including
static field initializers) of NixParserDefinition. The reason is that
NixParserDefinition is loaded very early during application startup
regardless of whether the language is used, and loading NixTypes could
cause noticable overhead. This warning caused me to initiate this
refactoring.